fixup! extract `tree_nodes`; add local `require`
